Refactor JobOfferCollectionViewCell and add associated assets for job listing display
- Now the component has a .state property with .normal, .filled and .open support
