+++ Freela-onTap/App/Assets.xcassets/person.crop.circle.fill.badge.magnifyingglass.imageset/Contents.json
@@ -1,0 +1,23 @@
+{
+  "images" : [
+    {
+      "filename" : "person.crop.circle.fill.badge.magnifyingglass.png",
+      "idiom" : "universal",
+      "scale" : "1x"
+    },
+    {
+      "filename" : "person.crop.circle.fill.badge.magnifyingglass@2x.png",
+      "idiom" : "universal",
+      "scale" : "2x"
+    },
+    {
+      "filename" : "person.crop.circle.fill.badge.magnifyingglass@3x.png",
+      "idiom" : "universal",
+      "scale" : "3x"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
